change id data type from int to long and make different tables for income and entries

+++ app/src/main/java/com/example/expensetracker/pojo/User.java
@@ -3,11 +3,11 @@ package com.example.expensetracker.pojo;
 import com.example.expensetracker.R;
 
 public class User extends SettingsParent {
-    private int id;
+    private long id;
      private String name;
      private int drawableId;
 
-    public User(int id, String name) {
+    public User(long id, String name) {
         this.id = id;
         this.name = name;
         this.drawableId = R.drawable.ic_launcher_background;
@@ -31,7 +31,7 @@ public class User extends SettingsParent {
         this.drawableId = drawableId;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
